fix: add providedIn:root to catchup service, fix blank favorites
