+++ libs/portal/shared/ui/src/lib/components/unified-collection/unified-live-tab.component.ts
@@ -13,10 +13,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { TranslatePipe } from '@ngx-translate/core';
-import {
-    isM3uCatchupPlaybackSupported,
-    resolveM3uCatchupUrl,
-} from '@iptvnator/shared/m3u-utils';
+import { isM3uCatchupPlaybackSupported } from '@iptvnator/shared/m3u-utils';
 import {
     DEFAULT_FAVORITES_CHANNEL_SORT_MODE,
     FavoritesChannelSortMode,
@@ -49,14 +46,13 @@ import {
     WebPlayerViewComponent,
 } from '@iptvnator/ui/playback';
 import { ResizableDirective } from '@iptvnator/ui/components';
-import { PlaylistsService, RuntimeCapabilitiesService, SettingsStore } from '@iptvnator/services';
-import { EpgItem, EpgProgram, Playlist } from '@iptvnator/shared/interfaces';
-import { XtreamUrlService } from '@iptvnator/portal/xtream/data-access';
+import { RuntimeCapabilitiesService, SettingsStore } from '@iptvnator/services';
+import { EpgItem, EpgProgram } from '@iptvnator/shared/interfaces';
 import {
     LiveEpgPanelComponent,
     LiveEpgPanelSummary,
 } from '@iptvnator/ui/shared-portals';
-import { firstValueFrom } from 'rxjs';
+import { UnifiedLiveCatchupService } from './unified-live-catchup.service';
 
 @Component({
     selector: 'app-unified-live-tab',
@@ -100,11 +96,7 @@ export class UnifiedLiveTabComponent {
     private readonly settingsStore = inject(SettingsStore);
     private readonly portalPlayer = inject(PORTAL_PLAYER);
     private readonly destroyRef = inject(DestroyRef);
-    private readonly xtreamUrlService = inject(XtreamUrlService);
-    private readonly playlistsService = inject(PlaylistsService);
-
-    /** Tracks the item whose detail is currently loaded for catchup lookups. */
-    private readonly activeItem = signal<UnifiedCollectionItem | null>(null);
+    readonly catchupService = inject(UnifiedLiveCatchupService);
 
     readonly player = this.settingsStore.player;
     readonly supportsEpg = this.runtime.supportsEpg;
@@ -155,31 +147,12 @@ export class UnifiedLiveTabComponent {
     );
     /** Portal EPG items converted to EpgProgram for the interactive list component. */
     readonly currentPortalEpgPrograms = computed<EpgProgram[]>(() =>
-        this.currentPortalEpgItems().map((item) => ({
-            start: item.start,
-            stop: item.stop ?? item.end,
-            channel: item.channel_id ?? item.id,
-            title: item.title,
-            desc: item.description ?? null,
-            category: null,
-            startTimestamp: this.parseEpgTimestamp(item.start_timestamp),
-            stopTimestamp: this.parseEpgTimestamp(item.stop_timestamp),
-        }))
+        this.catchupService.epgItemsToPrograms(this.currentPortalEpgItems())
     );
     /** Archive window (days) for the selected portal stream, or 0 if unavailable. */
-    readonly currentPortalArchiveDays = computed(() => {
-        const item = this.activeItem();
-        if (!item?.xtreamId) return 0;
-        // Match the all-channels detection: explicit 0 means no archive.
-        if (item.tvArchive === 0) return 0;
-        // Items that have a stored duration get that window; everything
-        // else (no tv_archive data in the DB) gets 0 so the EPG matches
-        // the all-channels view's "history but no catchup" notice.
-        if (item.tvArchiveDuration != null && item.tvArchiveDuration !== 0) {
-            return Math.max(0, Number(item.tvArchiveDuration));
-        }
-        return 0;
-    });
+    readonly currentPortalArchiveDays = computed(() =>
+        this.catchupService.portalArchiveDays(this.catchupService.activeItem())
+    );
     readonly activeRadioChannel = computed(() => {
         const channel = this.activeDetail()?.channel ?? null;
         return channel?.radio === 'true' ? channel : null;
@@ -347,94 +320,19 @@ export class UnifiedLiveTabComponent {
 
     async onProgramActivated(event: EpgProgramActivationEvent): Promise<void> {
         const detail = this.activeDetail();
-        const item = this.activeItem();
-        if (!detail) {
-            return;
-        }
+        const item = this.catchupService.activeItem();
+        if (!detail) return;
 
-        if (event.type === 'live') {
-            const originalDetail = item
-                ? await this.streamResolver.resolveLiveDetail(item)
-                : null;
-            if (originalDetail) {
-                this.activeDetail.set(originalDetail);
-            }
-            return;
-        }
+        const updatedDetail = await this.catchupService.onProgramActivated(
+            event,
+            detail,
+            item
+        );
+        if (!updatedDetail) return;
 
-        // Timeshift / catchup — M3U path
-        if (detail.epgMode === 'm3u' && detail.channel) {
-            const catchupUrl = resolveM3uCatchupUrl(
-                detail.channel,
-                event.program
-            );
-            if (catchupUrl) {
-                const updatedDetail = {
-                    ...detail,
-                    playback: {
-                        ...detail.playback,
-                        streamUrl: catchupUrl,
-                    },
-                };
-                this.activeDetail.set(updatedDetail);
-                if (this.shouldOpenExternalPlayback(updatedDetail, true)) {
-                    this.portalPlayer.openResolvedPlayback(
-                        updatedDetail.playback
-                    );
-                }
-            }
-            return;
-        }
-
-        // Timeshift / catchup — Xtream (portal) path
-        if (!item?.xtreamId) {
-            return;
-        }
-
-        try {
-            const credentials = await this.getXtreamCredentials(
-                item.playlistId
-            );
-            if (!credentials) {
-                return;
-            }
-
-            const startTimestamp = this.parseEpochSeconds(
-                event.program.startTimestamp,
-                event.program.start
-            );
-            const stopTimestamp = this.parseEpochSeconds(
-                event.program.stopTimestamp,
-                event.program.stop
-            );
-            if (startTimestamp == null || stopTimestamp == null) {
-                return;
-            }
-
-            const catchupUrl = await this.xtreamUrlService.resolveCatchupUrl(
-                item.playlistId,
-                credentials,
-                item.xtreamId,
-                startTimestamp,
-                stopTimestamp
-            );
-            if (catchupUrl) {
-                const updatedDetail = {
-                    ...detail,
-                    playback: {
-                        ...detail.playback,
-                        streamUrl: catchupUrl,
-                    },
-                };
-                this.activeDetail.set(updatedDetail);
-                if (this.shouldOpenExternalPlayback(updatedDetail, true)) {
-                    this.portalPlayer.openResolvedPlayback(
-                        updatedDetail.playback
-                    );
-                }
-            }
-        } catch {
-            // Keep current playback going if catchup fails.
+        this.activeDetail.set(updatedDetail);
+        if (this.shouldOpenExternalPlayback(updatedDetail, true)) {
+            this.portalPlayer.openResolvedPlayback(updatedDetail.playback);
         }
     }
 
@@ -443,7 +341,7 @@ export class UnifiedLiveTabComponent {
         this.isSelecting.set(false);
         this.activeDetail.set(null);
         this.activeUid.set(null);
-        this.activeItem.set(null);
+        this.catchupService.activeItem.set(null);
     }
 
     private async loadEpgMap(items: UnifiedCollectionItem[]): Promise<void> {
@@ -476,7 +374,7 @@ export class UnifiedLiveTabComponent {
         this.activeUid.set(item.uid);
         this.activeDetail.set(null);
         this.isSelecting.set(true);
-        this.activeItem.set(item);
+        this.catchupService.activeItem.set(item);
 
         try {
             const detail =
@@ -670,55 +568,4 @@ export class UnifiedLiveTabComponent {
         return Number.isFinite(parsedDate) ? parsedDate : null;
     }
 
-    private parseEpgTimestamp(
-        value: string | undefined
-    ): number | undefined {
-        const parsed = Number.parseInt(String(value ?? ''), 10);
-        return Number.isFinite(parsed) && parsed > 0 ? parsed : undefined;
-    }
-
-    private parseEpochSeconds(
-        timestamp: number | string | null | undefined,
-        fallbackIso: string
-    ): number | null {
-        const parsed = Number.parseInt(String(timestamp ?? ''), 10);
-        if (Number.isFinite(parsed) && parsed > 0) {
-            return parsed;
-        }
-        const ms = Date.parse(fallbackIso);
-        return Number.isFinite(ms) ? Math.floor(ms / 1000) : null;
-    }
-
-    private async getXtreamCredentials(
-        playlistId: string
-    ): Promise<{
-        serverUrl: string;
-        username: string;
-        password: string;
-    } | null> {
-        const electronPlaylist =
-            typeof window !== 'undefined'
-                ? await window.electron?.dbGetAppPlaylist?.(playlistId)
-                : null;
-        const playlist: Playlist | null =
-            electronPlaylist ??
-            (await firstValueFrom(
-                this.playlistsService.getPlaylistById(playlistId)
-            ));
-
-        if (
-            !playlist ||
-            !playlist.serverUrl ||
-            !playlist.username ||
-            !playlist.password
-        ) {
-            return null;
-        }
-
-        return {
-            serverUrl: playlist.serverUrl,
-            username: playlist.username,
-            password: playlist.password,
-        };
-    }
 }
